+++ CHANGELOG.md
@@ -7,10 +7,10 @@ Note that since we don't clearly distinguish between a public and private interf
 ## [Unreleased]
 
 - Add support for Glycam saccharide names
-- Add ``useInstanceGranularity`` option for marker, transparency, clipping, overpaint, substance data to save memory
+- Add ``instanceGranularity`` option for marker, transparency, clipping, overpaint, substance data to save memory
 - CellPack extension tweaks
     - Use instancing to create DNA/RNA curves to save memory
-    - Enable ``useInstanceGranularity`` by default
+    - Enable ``instanceGranularity`` by default
 
 ## [v3.9.1] - 2022-06-19